requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.10

# Security and authentication
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4

# File handling and environment
python-multipart==0.0.20
python-dotenv==1.0.0

# Data validation and utilities
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
python-dateutil==2.9.0.post0

# JWT handling
PyJWT